requirements.txt
python-dotenv
langchain
langchain_community
openai
flask
chainlit
transformers
huggingface_hub
PyPDF2
chroma
groq
ollama
bs4
unstructured
selenium
docx2txt
gitpython
arxiv
pymupdf
pinecone-client
langchain-pinecone
pinecone-notebooks
flask_pymongo
langchain_ollama
chromadb
langchain_openai
langchain_huggingface
pydantic
langchain-community
pymongo